Coordinate
Координатная сетка

+++ Server/nova-server/ship.h
@@ -83,6 +83,6 @@ public:
     CShip();
 };
 
-qint32 point_distance(QPoint* point1, QPoint* point2);
+qint64 point_distance(QPoint* point1, QPoint* point2);
 
 #endif // SHIP_H
